Fixed the RequestObject, NetworkThread, NetworkService, RequestDispatch. Database is setting and getting the response back to the fragments. Added the new layout for testing purpose.

+++ ActiveForecast/app/src/main/java/com/mike/givemewingzz/activeforecast/servermapping/WindData.java
@@ -9,7 +9,7 @@ public class WindData extends WeatherModel
 
     @Column(name = "WindDegree")
     @PropMap(serverFieldName = "deg")
-    public int windDegree;
+    public double windDegree;
 
     @Column(name = "WindSpeed")
     @PropMap(serverFieldName = "speed")
